feat: support for more browsers
